Revert AudioUrlPlayerService back to AudioUrlPlayer
