+++ common/src/main/java/io/homeassistant/companion/android/common/util/UtilModule.kt
@@ -18,4 +18,9 @@ object UtilModule {
     @Singleton
     fun provideAudioRecorder(@ApplicationContext appContext: Context): AudioRecorder =
         AudioRecorder(appContext.getSystemService<AudioManager>())
+
+    @Provides
+    @Singleton
+    fun provideAudioUrlPlayer(@ApplicationContext appContext: Context): AudioUrlPlayer =
+        AudioUrlPlayer(appContext.getSystemService<AudioManager>())
 }
