+++ wear/src/main/java/io/homeassistant/companion/android/conversation/ConversationViewModel.kt
@@ -15,6 +15,7 @@ import io.homeassistant.companion.android.common.data.prefs.WearPrefsRepository
 import io.homeassistant.companion.android.common.data.servers.ServerManager
 import io.homeassistant.companion.android.common.data.websocket.impl.entities.AssistPipelineResponse
 import io.homeassistant.companion.android.common.util.AudioRecorder
+import io.homeassistant.companion.android.common.util.AudioUrlPlayer
 import io.homeassistant.companion.android.conversation.views.AssistMessage
 import javax.inject.Inject
 import kotlinx.coroutines.launch
@@ -23,9 +24,10 @@ import kotlinx.coroutines.launch
 class ConversationViewModel @Inject constructor(
     private val serverManager: ServerManager,
     private val audioRecorder: AudioRecorder,
+    audioUrlPlayer: AudioUrlPlayer,
     private val wearPrefsRepository: WearPrefsRepository,
     application: Application
-) : AssistViewModelBase(serverManager, audioRecorder, application) {
+) : AssistViewModelBase(serverManager, audioRecorder, audioUrlPlayer, application) {
 
     companion object {
         private const val TAG = "ConvViewModel"
